+++ renderers/games/nohitter.py
@@ -3,7 +3,7 @@ try:
 except ImportError:
     from RGBMatrixEmulator import graphics
 
-import data.layout
+import data.config.layout as cfglayout
 import debug
 
 NOHITTER_TEXT = "N.H"
@@ -20,10 +20,10 @@ def render_nohit_text(canvas, layout, colors):
 
 
 def _get_nohitter_text(layout):
-    if layout.state == data.layout.LAYOUT_STATE_NOHIT:
+    if layout.state == cfglayout.LAYOUT_STATE_NOHIT:
         return NOHITTER_TEXT
 
-    if layout.state == data.layout.LAYOUT_STATE_PERFECT:
+    if layout.state == cfglayout.LAYOUT_STATE_PERFECT:
         return PERFECT_GAME_TEXT
 
     # If we get this far, the nohitter renderer probably shouldn't have been rendered.
